Refine=Optimize_Pose(fix 3D) + Optimize_3D_position(fix pose ,as BA)
The refinement can be seen as optimization problem where the variable we want to optimize is 3D position and Pose. So we create two function, one call optimize pose where it takes the neighbor pixel macthing to refine the pose, in this function we keep the 3D position fix. The second call optimize 3D point where we keep the pose fix but try to iterate to update the 3D position so that its projection is close to the gt query keypoint. The neighbor pixel take into consideration in projection space is 32